Undo last commit
Image centering problem due to translation from Yorick to Python. We have to correct ALL images at once to get correct results... Maybe this summer during sutra refactor, or you can search all the "+ 0.5" in shesha and replace it by "- 0.5" if it is relevant. Take care of things that are sent to GPU also...